typing and placement refactor
